+++ src/main/java/Model/Winner.java
@@ -11,7 +11,7 @@ public class Winner {
     public static List<String> findWinner(List<Car> carList) {
         int maxCount = findMaxCount(carList);
         carList.stream()
-                .filter(car -> car.isWinner(maxCount))
+                .filter(car -> isWinner(car,maxCount))
                 .forEach(x -> winnerList.add(x.carName.name));
         return winnerList;
     }
@@ -22,5 +22,7 @@ public class Winner {
                 .reduce(Integer::max)
                 .orElse(0);
     }
-
+    public static boolean isWinner(Car car,int maxCount) {
+        return maxCount == car.getPosition();
+    }
 }
