+++ src/main/java/Utils/StringToCar.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class StringToCar {
-    public List<Car> carList;
+    private final List<Car> carList;
     public StringToCar(List<String> carNames) {
         carList= new ArrayList<>();
         carNames.forEach(car -> carList.add(new Car(new CarName(car))));
