+++ src/test/java/Model/CarNameTest.java
@@ -8,7 +8,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class CarNameTest {
 
-    @DisplayName("이름이 널값이면 오류가 난다.")
+    @DisplayName("이름이 공백값이면 오류가 난다.")
     @Test
     void NullNameTest(){
         assertThatThrownBy(()->{
@@ -23,6 +23,4 @@ public class CarNameTest {
             CarName carName = new CarName("abcdef");
         }).isInstanceOf(IllegalArgumentException.class);
     }
-
-
 }
