+++ src/test/java/Model/CarTest.java
@@ -1,7 +1,6 @@
 package Model;
 
-import Model.Car;
-import Model.CarName;
+
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,14 +14,5 @@ public class CarTest {
         Car car = new Car(new CarName("abc"));
         car.move(number);
         Assertions.assertThat(car.position).isEqualTo(1);
-    }
-
-    @DisplayName("승리자가 맞으면 트루를 출력한다.")
-    @Test
-    void right_winner(){
-        int winnerPoint = 4;
-        Car car = new Car(new CarName("abc"));
-        car.position = 4;
-        Assertions.assertThat(car.isWinner(winnerPoint)).isEqualTo(true);
     }
 }
